pyproject.toml
[project]
name = "my-fastapi-nacos"
version = "0.1.1"
description = "FastAPI Nacos SDK"
readme = "README.md"
authors = [
    {name = "CarlosHuang", email = "dragon_hht@163.com"}
]
requires-python = ">=3.10"
dependencies = [
    "nacos-sdk-python==2.0.9",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.2",
    "uvicorn>=0.40.0",
    "loguru>=0.7.3",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "dataclasses>=0.8",
]

[tool.setuptools]
packages = [
    "my_fastapi_nacos",
    "my_fastapi_nacos.config",
    "my_fastapi_nacos.core",
    "my_fastapi_nacos.http",
    "my_fastapi_nacos.models",
    "my_fastapi_nacos.utils"
]
package-dir = {"" = "."}

[tool.pytest.ini_options]
pythonpath = ""
testpaths = ["tests"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
